+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.20
 
 -   Fix: [FluffOS `ref` keyword not being parsed correctly. #155](https://github.com/jlchmura/lpc-language-server/issues/155)
+-   Fix: [Object type not resolved when used in a union #156](https://github.com/jlchmura/lpc-language-server/issues/156)
 -   `null` keyword was incorrectly parsed as a reserved word in LPC code.
 
 ## 1.1.19